Autonomous-Data-Analyst-Agent/requirements.txt
# Core
pandas>=2.1.0,<2.2
numpy>=1.25.0,<2.2    # compatible with numba
python-dotenv>=1.0.0
requests>=2.31.0
regex>=2023.8.8

# NLP
nltk>=3.8.1
spacy>=3.6.1
wordcloud>=1.8.2.2
langdetect>=1.0.9

# File loaders
unstructured>=0.9.12
PyPDF2>=3.0.1
python-docx>=0.8.11
openpyxl>=3.1.2

# LangChain + embeddings
langchain>=0.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Optional for images
pillow>=10.0.0,<11    # compatible with streamlit

# Optional visualization
matplotlib>=3.8.0
seaborn>=0.12.2
plotly>=6.0.0

# Optional for unstructured inference
onnxruntime>=1.18.0    # required by unstructured-inference >=1.0.5
